+++ .idea/sqldialects.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/AccountService/src/main/resources/data-test.sql" dialect="H2" />
-    <file url="file://$PROJECT_DIR$/AccountService/src/main/resources/schema.sql" dialect="H2" />
+    <file url="file://$PROJECT_DIR$/account-service/src/main/resources/data-test.sql" dialect="H2" />
+    <file url="file://$PROJECT_DIR$/account-service/src/main/resources/schema.sql" dialect="H2" />
     <file url="PROJECT" dialect="H2" />
   </component>
 </project>
