--- App/src/main/java/com/project/chat/ChatApplication.java
@@ -1,13 +1,0 @@
-package com.project.chat;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ChatApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ChatApplication.class, args);
-	}
-
-}
