--- App/src/test/java/com/project/chat/ChatApplicationTests.java
@@ -1,13 +1,0 @@
-package com.project.chat;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ChatApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
