--- EurekaServer/src/main/resources/application.properties
@@ -1,6 +1,0 @@
-spring.application.name=EurekaServer
-server.port=8761
-eureka.client.register-with-eureka=false
-eureka.client.fetch-registry=false
-logging.level.com.netflix.eureka=OFF
-logging.level.com.netflix.discovery=OFF
