--- EurekaServer/src/test/java/com/project/eurekaServer/EurekaServerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.project.eurekaServer;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EurekaServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
