+++ eureka-server/src/test/java/com/project/eurekaserver/EurekaServerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.project.eurekaserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EurekaServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
