+++ gate-way/src/main/java/com/project/gateway/GateWayApplication.java
@@ -1,0 +1,26 @@
+package com.project.gateway;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.servlet.function.RouterFunction;
+import org.springframework.web.servlet.function.ServerResponse;
+
+
+import static org.springframework.cloud.gateway.server.mvc.handler.GatewayRouterFunctions.route;
+import static org.springframework.cloud.gateway.server.mvc.handler.HandlerFunctions.http;
+
+@SpringBootApplication
+public class GateWayApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(GateWayApplication.class, args);
+
+	}
+	
+	@Bean
+	public RouterFunction<ServerResponse> getRoute() {
+		return route().GET("/get",http("http://localhost:8761/")).build();
+	}
+
+}
