+++ gate-way/src/test/java/com/project/gateway/GateWayApplicationTests.java
@@ -1,0 +1,13 @@
+package com.project.gateway;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GateWayApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
